+++ src/app/initializer.ts
@@ -2,8 +2,11 @@ import { CuiSetupInit } from "../core/models/setup";
 import { is } from "../core/utlis/functions";
 import { CuiInstance } from "./instance";
 import { ICONS } from "../core/utlis/statics";
-import { ICuiPlugin } from "../core/models/interfaces";
+import { ICuiPlugin, CuiInitData, CuiInitResult, ICuiComponent } from "../core/models/interfaces";
 import { CuiAutoLightModePlugin } from "./plugins/light/light";
+import { CuiIconComponent } from "../components/icon/icon";
+import { CuiCircleComponent } from "../components/circle/circle";
+import { CuiSpinnerComponent } from "../components/spinner/spinner";
 
 export class CuiInitializer {
     #window: any;
@@ -11,22 +14,31 @@ export class CuiInitializer {
         this.#window = window;
     }
 
-    async init(plugins: ICuiPlugin[], setup?: CuiSetupInit): Promise<boolean> {
-        const settings: CuiSetupInit = { ... new CuiSetupInit(), ...setup }
+    async init(setup: CuiInitData): Promise<CuiInitResult> {
+        let settings: CuiSetupInit = { ... new CuiSetupInit(), ...setup.setup }
         const appPrefix: string = settings.app;
+        const result: CuiInitResult = {
+            result: false
+        }
         if (is(this.#window[appPrefix])) {
-            return false;
+            result.message = "Instance is already initialized";
+            return result;
         }
-        this.#window[appPrefix] = new CuiInstance(settings, plugins)
-        this.#window[appPrefix].init();
-        return true;
-    }
-
-    async setIcons(icons: any): Promise<boolean> {
-        for (let icon in icons) {
-            ICONS[icon] = icons[icon];
+        if (is(setup.icons)) {
+            for (let icon in setup.icons) {
+                ICONS[icon] = setup.icons[icon];
+            }
         }
-        return true;
+        try {
+            this.#window[appPrefix] = new CuiInstance(settings, setup.plugins, setup.components)
+            this.#window[appPrefix].init();
+        } catch (e) {
+            console.error(e);
+            result.message = "An error occured during initialization";
+            return result;
+        }
+        result.result = true;
+        return result;
     }
 }
 
@@ -46,12 +58,23 @@ export class CuiInit {
             new CuiAutoLightModePlugin({ autoLight: true })
         ];
 
-        if (is(icons)) {
-            initializer.setIcons(icons)
-        }
-        if (initializer.init(plugins, setup)) {
+        const components: ICuiComponent[] = [
+            new CuiIconComponent(),
+            new CuiCircleComponent(),
+            new CuiSpinnerComponent()
+        ];
+
+        let result = await initializer.init({
+            setup: setup,
+            icons: icons,
+            plugins: plugins,
+            components: components
+        })
+        if (result.result) {
             this.#isInitialized = true;
             return true;
+        } else {
+            console.error(`A cUI instance failed to initialize: [${result.message ?? "#"}]`);
         }
         console.log("Cui Light failed to init")
         return false;