+++ src/core/bus/bus.ts
@@ -1,5 +1,8 @@
-import { ICuiCallbackExecutor, ICuiEventBus, CuiEventReceiver, ICuiLogger, ICuiEventEmitHandler, CuiContext } from "../models/interfaces";
-import { is } from "../utlis/functions";
+import { ICuiEventBus, CuiEventReceiver, ICuiLogger, ICuiEventEmitHandler, CuiContext } from "../models/interfaces";
+import { is, are } from "../utlis/functions";
+import { ArgumentError, CuiBusError } from "../models/errors";
+import { CuiLogger } from "../utlis/logger";
+import { CuiLoggerFactory } from "../factories/logger";
 
 export class CuiEventBus implements ICuiEventBus {
     #events: { [event: string]: CuiEventReceiver }
@@ -9,12 +12,16 @@ export class CuiEventBus implements ICuiEventBus {
     constructor(emitHandler: ICuiEventEmitHandler) {
         this.#events = {};
         this.#eventHandler = emitHandler;
+        this.#log = CuiLoggerFactory.get("CuiEventBus");
     }
 
-    on(name: string, callback: CuiContext, ctx: CuiContext): void {
+    on(name: string, callback: any, ctx: CuiContext): void {
+        if (!are(name, callback, ctx)) {
+            throw new ArgumentError("Missing argument")
+        }
         let id = ctx.getCuid();
         if (!is(id)) {
-            throw new Error("Missing component name or cuid")
+            throw new CuiBusError("Missing component name or cuid")
         }
         this.#log.debug(`Attaching new event: [${name}] for: [${id}]`)
         if (!this.#events[name]) {
@@ -25,6 +32,9 @@ export class CuiEventBus implements ICuiEventBus {
     }
 
     detach(name: string, ctx: CuiContext): void {
+        if (!are(name, ctx)) {
+            throw new ArgumentError("Missing argument")
+        }
         let ev = this.#events[name]
         let id = ctx.getCuid();
         this.#log.debug(`Detaching item: [${id}] from [${name}]`)
@@ -34,12 +44,17 @@ export class CuiEventBus implements ICuiEventBus {
     }
 
     detachAll(name: string): void {
-        if (name && this.#events[name]) {
+        if (is(name) && this.#events[name]) {
             delete this.#events[name]
+        } else {
+            this.#log.error(`Event name is missing or incorrect`, "detachAll")
         }
     }
 
     async emit(event: string, ...args: any[]): Promise<boolean> {
+        if (!is(event)) {
+            throw new ArgumentError("Event name is incorrect");
+        }
         this.#log.debug(`Emit: [${event}]`)
         await this.#eventHandler.handle(this.#events[event], args)
         return true;
@@ -48,13 +63,6 @@ export class CuiEventBus implements ICuiEventBus {
     isSubscribing(name: string, ctx: CuiContext) {
         let ev = this.#events[name]
         return this.isAttached(ev, ctx.getCuid())
-    }
-
-    clear(name: string): void {
-        let ev = this.#events[name];
-        if (ev) {
-            delete this.#events[name]
-        }
     }
 
     private isAttached(ev: CuiEventReceiver, id: string): boolean {