+++ src/core/utlis/types.ts
@@ -2,4 +2,4 @@ export type CuiLogLevel = "none" | "error" | "warning" | "debug";
 export type CuiInteractionsType = 'sync' | 'async';
 export type CuiColorSetType = 'light' | 'dark' | 'accent' | 'secondary' | 'success' | 'warning' | 'error';
 export type CuiClearCacheType = 'element' | "collection" | "all";
-export type CuiLightMode = 'light' | 'dark'
+export type CuiLightMode = 'light' | 'dark';