+++ tests/helpers/models.ts
@@ -1,4 +1,4 @@
-import { CuiCachable } from "../../src/core/models/interfaces";
+import { CuiCachable, CuiContext } from "../../src/core/models/interfaces";
 
 export class CacheTestItem implements CuiCachable {
     id: number;
@@ -7,5 +7,32 @@ export class CacheTestItem implements CuiCachable {
     }
     refresh(): boolean {
         return true;
+    }
+}
+
+export class ExecutorTestItem {
+    value: boolean;
+    constructor() {
+        this.value = false;
+    }
+
+    setValue(value: boolean) {
+        this.value = value;
+    }
+}
+
+export class ExecutorTestItemExt implements CuiContext {
+    value: boolean;
+    id: string;
+    constructor(id: string) {
+        this.value = false;
+        this.id = id;
+    }
+    getCuid(): string {
+        return this.id;
+    }
+
+    setValue(value: boolean) {
+        this.value = value;
     }
 }